refactor(bub-qq): split package into protocol, gateway, inbound, and outbound layers
Keep the plugin surface the same while grouping QQ protocol, receive transports, and C2C IO. Also allow an empty receive_mode so bub onboard can validate an unconfigured qq section.
