+++ packages/bub-qq/src/bub_qq/gateway/ws_errors.py
