+++ packages/bub-qq/src/bub_qq/inbound/__init__.py
@@ -1,0 +1,13 @@
+"""Inbound event adaptation into Bub ChannelMessage."""
+
+from .c2c import QQC2CDeduper
+from .c2c import QQC2CInboundService
+from .c2c import QQC2CSessionState
+from .c2c import build_c2c_channel_message
+
+__all__ = [
+    "QQC2CDeduper",
+    "QQC2CInboundService",
+    "QQC2CSessionState",
+    "build_c2c_channel_message",
+]
