+++ packages/bub-qq/src/bub_qq/outbound/__init__.py
@@ -1,0 +1,7 @@
+"""Outbound delivery to QQ OpenAPI."""
+
+from .c2c import QQC2CSendService
+
+__all__ = [
+    "QQC2CSendService",
+]
