+++ packages/bub-qq/src/bub_qq/protocol/__init__.py
@@ -1,0 +1,17 @@
+"""QQ Open Platform protocol: auth, OpenAPI client, signatures, and models."""
+
+from .auth import QQAuthError
+from .auth import QQTokenProvider
+from .errors import QQKnownOpenAPIError
+from .errors import QQOpenAPIError
+from .models import QQC2CMessage
+from .openapi import QQOpenAPI
+
+__all__ = [
+    "QQAuthError",
+    "QQC2CMessage",
+    "QQKnownOpenAPIError",
+    "QQOpenAPI",
+    "QQOpenAPIError",
+    "QQTokenProvider",
+]
