+++ packages/bub-qq/src/bub_qq/protocol/signature.py
